Moved method assert_type_mml closer to its use in namespace app.schema
