+++ source/app/util.py
@@ -23,7 +23,6 @@ import base64
 import datetime
 import hashlib
 import logging as log
-import marshmallow
 import shutil
 import weakref
 from cryptography.exceptions import InvalidSignature
@@ -150,42 +149,3 @@ def str_to_bool(value):
         return bool(value)
 
     return value.lower() in ['true', '1', 'yes', 'y', 't']
-
-
-def assert_type_mml(input_var: any, field_name: str, type: type, allow_none: bool = False,
-                    max_len: int = None, max_val: int = None, min_val: int = None):
-    if input_var is None:
-        if allow_none is False:
-            raise marshmallow.ValidationError('Invalid data - non null expected',
-                                              field_name=field_name if field_name else 'type')
-        return True
-
-    if isinstance(input_var, type):
-        if max_len:
-            if len(input_var) > max_len:
-                raise marshmallow.ValidationError('Invalid data - max length exceeded',
-                                                  field_name=field_name if field_name else 'type')
-
-        if max_val:
-            if input_var > max_val:
-                raise marshmallow.ValidationError('Invalid data - max value exceeded',
-                                                  field_name=field_name if field_name else 'type')
-
-        if min_val:
-            if input_var < min_val:
-                raise marshmallow.ValidationError('Invalid data - min value exceeded',
-                                                  field_name=field_name if field_name else 'type')
-
-        return True
-
-    try:
-
-        if isinstance(type(input_var), type):
-            return True
-
-    except Exception as e:
-        log.error(e)
-        print(e)
-
-    raise marshmallow.ValidationError('Invalid data type',
-                                      field_name=field_name if field_name else 'type')
